renamed the stopwatch library files to libstopwatch and only initialize the HRESULT hr variable once in the LoadWaveAudioFile method
