+++ directx_window/directx_window.vcxproj
@@ -131,7 +131,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="directx_window.cpp" />
-    <ClCompile Include="stopwatch.cpp" />
+    <ClCompile Include="libstopwatch.cpp" />
     <ClCompile Include="xaudio_driver.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -142,7 +142,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="resource.h" />
-    <ClInclude Include="stopwatch.h" />
+    <ClInclude Include="libstopwatch.h" />
     <ClInclude Include="xaudio.h" />
     <ClInclude Include="xaudio_driver.h" />
   </ItemGroup>
