+++ directx_window/libstopwatch.h
